+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This file is a history of the changes made to idearium-lib.
 
-## Unreleased
+## v1.0.0-alpha.34 (25 May 2018)
 
 - Updated apm to allow custom `unhandledRejection` and `apm.handleUncaughtExceptions` exception handlers. Just set `apm.exception = () => { ... }` in the app.
 - Updated the default exception handler to call `log.error` instead of just console.log.